Split multi.jl, Move distributed processing code into its own module. Incorporate changes due to #20376
