+++ base/parallel/pmap.jl
@@ -20,10 +20,10 @@ for details.
 """
 function pgenerate(p::WorkerPool, f, c)
     if length(p) == 0
-        return AsyncGenerator(f, c; ntasks=()->nworkers(p))
+        return Base.AsyncGenerator(f, c; ntasks=()->nworkers(p))
     end
     batches = batchsplit(c, min_batch_count = length(p) * 3)
-    return Iterators.flatten(AsyncGenerator(remote(p, b -> asyncmap(f, b)), batches))
+    return Iterators.flatten(Base.AsyncGenerator(remote(p, b -> asyncmap(f, b)), batches))
 end
 pgenerate(p::WorkerPool, f, c1, c...) = pgenerate(p, a->f(a...), zip(c1, c...))
 pgenerate(f, c) = pgenerate(default_worker_pool(), f, c)
