+++ base/precompile.jl
@@ -107,7 +107,7 @@ precompile(Base.LineEdit.write_prompt, (Base.Terminals.TTYTerminal, Base.LineEdi
 precompile(Base.LineEdit.write_prompt, (Base.Terminals.TerminalBuffer, Base.LineEdit.PromptState))
 precompile(Base.Multimedia.TextDisplay, (Base.TTY,))
 precompile(Base.Multimedia.display, (Int,))
-precompile(Base.ProcessGroup, (Int, Array{Any,1}, Array{Any,1}))
+precompile(Base.Parallel.ProcessGroup, (Int, Array{Any,1}, Array{Any,1}))
 precompile(Base.REPL.:(==), (Base.REPL.REPLDisplay{Base.REPL.LineEditREPL}, Base.REPL.REPLDisplay{Base.REPL.LineEditREPL}))
 precompile(Base.REPL.LineEditREPL, (Base.Terminals.TTYTerminal, Bool, String, String, String, String, String, Bool, Bool, Bool, Bool))
 precompile(Base.REPL.LineEditREPL, (Base.Terminals.TTYTerminal,))
@@ -140,8 +140,8 @@ precompile(Base.REPLCompletions.completions, (String, Int))
 precompile(Base.Random.srand, ())
 precompile(Base.Random.srand, (String, Int))
 precompile(Base.Random.srand, (UInt,))
-precompile(Base.RemoteChannel, (Int, Int, Int))
-precompile(Base.RemoteValue, ())
+precompile(Base.Parallel.RemoteChannel, (Int, Int, Int))
+precompile(Base.Parallel.RemoteValue, ())
 precompile(Base.Set, ())
 precompile(Base.SystemError, (String,))
 precompile(Base.TCPSocket, (Ptr{Void},))
@@ -207,8 +207,8 @@ precompile(Base.fdio, (Int32,))
 precompile(Base.fill!, (BitArray{1}, Bool))
 precompile(Base.first, (UnitRange{Int},))
 precompile(Base.flush, (IOStream,))
-precompile(Base.flush_gc_msgs, ())
-precompile(Base.flush_gc_msgs, (Base.Worker,))
+precompile(Base.Parallel.flush_gc_msgs, ())
+precompile(Base.Parallel.flush_gc_msgs, (Base.Parallel.Worker,))
 precompile(Base.get, (Base.EnvHash, String, String))
 precompile(Base.get, (Dict{Any, Any}, Tuple{Int, Int}, Bool))
 precompile(Base.get, (Dict{Any,Any}, Symbol, String))
@@ -223,8 +223,8 @@ precompile(Base.getindex, (Type{String}, String, String))
 precompile(Base.getindex, (Type{Dict{Any, Any}}, Dict{Any, Any}, Dict{Any, Any}, Dict{Any, Any}, Dict{Any, Any}, Dict{Any, Any}))
 precompile(Base.getpid, ())
 precompile(Base.hash, (Int,))
-precompile(Base.hash, (RemoteChannel, UInt))
-precompile(Base.hash, (RemoteChannel,))
+precompile(Base.hash, (Base.Parallel.RemoteChannel, UInt))
+precompile(Base.hash, (Base.Parallel.RemoteChannel,))
 precompile(Base.haskey, (Base.EnvHash, String))
 precompile(Base.haskey, (Dict{Symbol,Any}, Symbol))
 precompile(Base.haskey, (ObjectIdDict, Symbol))
@@ -254,8 +254,8 @@ precompile(Base.isequal, (Base.LineEdit.Prompt, Base.LineEdit.Prompt))
 precompile(Base.isequal, (Bool, Bool))
 precompile(Base.isequal, (Char, String))
 precompile(Base.isequal, (Int,Int))
-precompile(Base.isequal, (RemoteChannel, RemoteChannel))
-precompile(Base.isequal, (RemoteChannel, WeakRef))
+precompile(Base.isequal, (Base.Parallel.RemoteChannel, Base.Parallel.RemoteChannel))
+precompile(Base.isequal, (Base.Parallel.RemoteChannel, WeakRef))
 precompile(Base.isequal, (Symbol, Symbol))
 precompile(Base.isequal, (VersionNumber, VersionNumber))
 precompile(Base.isequal, (Void, Void))
@@ -333,7 +333,7 @@ precompile(Base.reinit_stdio, ())
 precompile(Base.repeat, (String, Int))
 precompile(Base.repl_cmd, (Cmd, Base.Terminals.TTYTerminal))
 precompile(Base.require, (Symbol,))
-precompile(Base.remoteref_id, (RemoteChannel,))
+precompile(Base.Parallel.remoteref_id, (Base.Parallel.RemoteChannel,))
 precompile(Base.rsearch, (String, Char))
 precompile(Base.rstrip, (String,))
 precompile(Base.run, (Cmd,))
@@ -388,12 +388,12 @@ precompile(Base.Symbol, (SubString{String},))
 precompile(Base.sync_begin, ())
 precompile(Base.sync_end, ())
 precompile(Base.systemerror, (Symbol, Bool))
-precompile(Base.take!, (Base.RemoteValue,))
-precompile(Base.take!, (RemoteChannel,))
-precompile(Base.take_ref, (Tuple{Int,Int},))
+precompile(Base.Parallel.take!, (Base.Parallel.RemoteValue,))
+precompile(Base.Parallel.take!, (Base.Parallel.RemoteChannel,))
+precompile(Base.Parallel.take_ref, (Tuple{Int,Int},))
 precompile(Base.take!, (IOBuffer,))
 precompile(Base.task_local_storage, ())
-precompile(Base.terminate_all_workers, ())
+precompile(Base.Parallel.terminate_all_workers, ())
 precompile(Base.try_include, (String,))
 precompile(Base.UInt, (UInt,))
 precompile(Base.unsafe_copy!, (Array{Dict{Any, Any}, 1}, Int, Array{Dict{Any, Any}, 1}, Int, Int))
@@ -403,7 +403,7 @@ precompile(Base.uv_error, (String, Bool))
 precompile(Base.uvfinalize, (Base.TTY,))
 precompile(Base.vcat, (Base.LineEdit.Prompt,))
 precompile(Base.wait, ())
-precompile(Base.wait, (RemoteChannel,))
+precompile(Base.Parallel.wait, (Base.Parallel.RemoteChannel,))
 precompile(Base.write, (Base.Terminals.TTYTerminal, String))
 precompile(Base.write, (Base.Terminals.TerminalBuffer, String))
 precompile(Base.write, (IOBuffer, Vector{UInt8}))
@@ -441,14 +441,13 @@ precompile(Base.similar, (Array{Base.LineEdit.Prompt, 1}, Type{Base.LineEdit.Tex
 precompile(Base.setindex!, (Array{Base.LineEdit.TextInterface, 1}, Base.LineEdit.PrefixHistoryPrompt{Base.REPL.REPLHistoryProvider}, Base.UnitRange{Int64}))
 precompile(Base.LineEdit.init_state, (Base.Terminals.TTYTerminal, Base.LineEdit.PrefixHistoryPrompt{Base.REPL.REPLHistoryProvider}))
 precompile(Base.setindex!, (Base.Dict{Any, Any}, Base.LineEdit.PrefixSearchState, Base.LineEdit.PrefixHistoryPrompt{Base.REPL.REPLHistoryProvider}))
-precompile(Base.take_ref, (Tuple{Int64, Int64},))
 precompile(Base.get, (Base.Dict{Any, Any}, Tuple{Int64, Int64}, Bool))
 precompile(Base.LineEdit.refresh_multi_line, (Array{Any, 1}, Base.Terminals.TerminalBuffer, Base.Terminals.TTYTerminal, Base.IOBuffer, Base.LineEdit.InputAreaState, Base.LineEdit.PromptState))
 precompile(Base.schedule, (Array{Any, 1}, Task, Void))
 precompile(Base.LineEdit.match_input, (Function, Base.LineEdit.MIState, Base.Terminals.TTYTerminal, Array{Char, 1}, Base.Dict{Char, Any}))
-precompile(==, (Base.RemoteChannel, WeakRef))
-precompile(==, (Base.RemoteChannel, Base.RemoteChannel))
-precompile(Base.send_del_client, (Base.RemoteChannel,))
+precompile(==, (Base.Parallel.RemoteChannel, WeakRef))
+precompile(==, (Base.Parallel.RemoteChannel, Base.Parallel.RemoteChannel))
+precompile(Base.Parallel.send_del_client, (Base.Parallel.RemoteChannel,))
 precompile(!=, (Base.SubString{String}, String))
 precompile(Base.join, (Base.IOBuffer, Array{Base.SubString{String}, 1}, String))
 precompile(Base.joinpath, (String, String, String, String))
@@ -460,22 +459,22 @@ precompile(Base.vect, (Base.LineEdit.Prompt, String))
 isdefined(Core, :Inference) && Base.code_typed(Base.code_typed)
 
 # Speeding up addprocs for LocalManager
-precompile(Base.start_worker, ())
-precompile(Base.start_worker, (Base.TTY,))
-precompile(Base.process_messages, (Base.TCPSocket, Base.TCPSocket))
-precompile(Base.process_messages, (Base.TCPSocket, Base.TCPSocket, Void))
-precompile(Base.process_tcp_streams, (Base.TCPSocket, Base.TCPSocket, Void))
-precompile(Base.message_handler_loop, (Base.TCPSocket, Base.TCPSocket, Void))
-precompile(Base.connect_to_peer, (Base.LocalManager, Int64, Base.WorkerConfig))
-precompile(Base.connect, (Base.LocalManager, Int64, Base.WorkerConfig))
-precompile(Base.connect_w2w, (Int64, Base.WorkerConfig))
-precompile(Base.connect_to_worker, (String, Int64))
-precompile(Base.addprocs, (Base.LocalManager, ))
-precompile(Base.addprocs, (Int, ))
-precompile(Base.setup_launched_worker, (Base.LocalManager, Dict, Base.WorkerConfig, Array{Int,1}))
-precompile(Base.create_worker, (Base.LocalManager, Dict, Base.WorkerConfig))
-precompile(Base.launch, (Base.LocalManager, Dict, Array{Base.WorkerConfig, 1}, Base.Condition))
-precompile(Base.set_valid_processes, (Array{Int, 1}, ))
+precompile(Base.Parallel.start_worker, ())
+precompile(Base.Parallel.start_worker, (Base.TTY,))
+precompile(Base.Parallel.process_messages, (Base.TCPSocket, Base.TCPSocket))
+precompile(Base.Parallel.process_messages, (Base.TCPSocket, Base.TCPSocket, Void))
+precompile(Base.Parallel.process_tcp_streams, (Base.TCPSocket, Base.TCPSocket, Void))
+precompile(Base.Parallel.message_handler_loop, (Base.TCPSocket, Base.TCPSocket, Void))
+precompile(Base.Parallel.connect_to_peer, (Base.Parallel.LocalManager, Int64, Base.Parallel.WorkerConfig))
+precompile(Base.Parallel.connect, (Base.Parallel.LocalManager, Int64, Base.Parallel.WorkerConfig))
+precompile(Base.Parallel.connect_w2w, (Int64, Base.Parallel.WorkerConfig))
+precompile(Base.Parallel.connect_to_worker, (String, Int64))
+precompile(Base.Parallel.addprocs, (Base.Parallel.LocalManager, ))
+precompile(Base.Parallel.addprocs, (Int, ))
+precompile(Base.Parallel.setup_launched_worker, (Base.Parallel.LocalManager, Dict, Base.Parallel.WorkerConfig, Array{Int,1}))
+precompile(Base.Parallel.create_worker, (Base.Parallel.LocalManager, Dict, Base.Parallel.WorkerConfig))
+precompile(Base.Parallel.launch, (Base.Parallel.LocalManager, Dict, Array{Base.Parallel.WorkerConfig, 1}, Base.Condition))
+precompile(Base.Parallel.set_valid_processes, (Array{Int, 1}, ))
 
 # Speed up repl help
 sprint(Markdown.term, @doc mean)
