+++ base/sysimg.jl
@@ -283,10 +283,8 @@ importall .Enums
 include("serialize.jl")
 importall .Serializer
 include("channels.jl")
-include("clusterserialize.jl")
-include("multi.jl")
-include("workerpool.jl")
-include("managers.jl")
+include("parallel/Parallel.jl")
+importall .Parallel
 
 # code loading
 include("loading.jl")
@@ -362,9 +360,7 @@ import .Dates: Date, DateTime, DateFormat, @dateformat_str, now
 include("sparse/sparse.jl")
 importall .SparseArrays
 
-# parallel map
 include("asyncmap.jl")
-include("pmap.jl")
 
 # worker threads
 include("threadcall.jl")
@@ -388,7 +384,7 @@ function __init__()
     Multimedia.reinit_displays() # since Multimedia.displays uses STDOUT as fallback
     early_init()
     init_load_path()
-    init_parallel()
+    Parallel.init_parallel()
     init_threadcall()
 end
